refactor(import): altera modules em ProjectList
