+++ src/components/project-input.ts
@@ -1,4 +1,4 @@
-import { Component } from "../components/base-component.js";
+import { Component } from "./base-component.js";
 import { Autobind } from "../decorators/autobind.js";
 import { validate, Validatable } from "../util/validation.js";
 import { projetoState } from "../state/project-state";
