+++ src/components/project-item.ts
@@ -1,4 +1,4 @@
-import { Component } from "../components/base-component.js";
+import { Component } from "./base-component.js";
 import { Project } from "../models/project.js";
 import { Autobind } from "../decorators/autobind.js";
 import { Draggable } from "../models/drag-drop.js";
